Search in Library sections
Add a search bar for Bookmarks, History, and Downloads.

Each view keeps a cached version of the list of items. When the
user enters a search term, we filter this cached list and update
the Adapter. This avoids having to query the backend each time.

Currently this implementation updates the search results as the
user types each character, but if this does not have good performance
we can change it later on.

There is a change to the OnScrollListener to ensure that the
RecyclerView only requests the focus when it scrolls as a result
of user interaction, which will close the on-screen keyboard.